+++ Engine/TextureShader.h
@@ -15,14 +15,14 @@ public:
 	TextureShader( TextureShader&& donor ) = delete;
 	TextureShader& operator=( TextureShader&& donor ) = delete;
 
-	bool Render( ID3D11DeviceContext* pDeviceContext, UINT nVertices, ID3D11ShaderResourceView* pTexture );
-	bool RenderIndexed( ID3D11DeviceContext* pDeviceContext, UINT nIndexes, ID3D11ShaderResourceView* pTexture );
+	bool Render( ID3D11DeviceContext* pDeviceContext, size_t nVertices, ID3D11ShaderResourceView* pTexture );
+	bool RenderIndexed( ID3D11DeviceContext* pDeviceContext, size_t nIndexes, ID3D11ShaderResourceView* pTexture );
 private:
 	void OutputShaderErrorMessage( ID3DBlob* errorMessage, HWND hWnd, const std::wstring& shaderFilename );
 	void SetShaderParameters( ID3D11DeviceContext* pDeviceContext, ID3D11ShaderResourceView* pSrv );
 
-	void RenderShader( ID3D11DeviceContext* pDeviceContext, UINT nVertices );
-	void RenderShaderIndexed( ID3D11DeviceContext* pDeviceContext, UINT nIndexes );
+	void RenderShader( ID3D11DeviceContext* pDeviceContext, size_t nVertices );
+	void RenderShaderIndexed( ID3D11DeviceContext* pDeviceContext, size_t nIndexes );
 private:
 	Microsoft::WRL::ComPtr<ID3D11VertexShader>	m_pVertexShader;
 	Microsoft::WRL::ComPtr<ID3D11PixelShader>	m_pPixelShader;
