+++ Engine/VertexBuffer.h
@@ -11,7 +11,7 @@ public:
 	{
 		D3D11_BUFFER_DESC vertexBufferDesc;
 		vertexBufferDesc.Usage = D3D11_USAGE_DEFAULT;
-		vertexBufferDesc.ByteWidth = sizeof( V ) * vertices.size();
+		vertexBufferDesc.ByteWidth = UINT( sizeof( V ) * vertices.size() );
 		vertexBufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 		vertexBufferDesc.CPUAccessFlags = 0;
 		vertexBufferDesc.MiscFlags = 0;
